+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - Enh #310: Add JSON overlaps condition builder (@Tigrov)
 - Enh #312: Update `bit` type according to main PR yiisoft/db#860 (@Tigrov)
 - Enh #315: Raise minimum PHP version to `^8.1` with minor refactoring (@Tigrov)
+- Enh #314: Implement `ColumnFactory` class (@Tigrov)
 
 ## 1.2.0 March 21, 2024